+++ src/proccontrol/pcontrol_mutatee_tools.h
@@ -51,7 +51,7 @@ int releaseThreads();
 int send_message(unsigned char *msg, size_t msg_size);
 int recv_message(unsigned char *msg, size_t msg_size);
 
-uint64_t getFunctionPtr(unsigned long *ptr);
+intptr_t getFunctionPtr(intptr_t *ptr);
 uint64_t getTOCValue(unsigned long *ptr);
 
 #endif
